Skilt ut klassene i Journalpost.kt til egne filer. Lagt til JournalposterForBrukerRequest.kt
