+++ felles/src/main/kotlin/no/nav/familie/kontrakter/felles/journalpost/Bruker.kt
@@ -1,0 +1,4 @@
+package no.nav.familie.kontrakter.felles.journalpost
+
+data class Bruker(val id: String,
+                  val type: BrukerIdType)
