+++ felles/src/main/kotlin/no/nav/familie/kontrakter/felles/journalpost/BrukerIdType.kt
@@ -1,0 +1,7 @@
+package no.nav.familie.kontrakter.felles.journalpost
+
+enum class BrukerIdType {
+    AKTOERID,
+    FNR,
+    ORGNR
+}
