+++ felles/src/main/kotlin/no/nav/familie/kontrakter/felles/journalpost/Dokumentstatus.kt
@@ -1,0 +1,8 @@
+package no.nav.familie.kontrakter.felles.journalpost
+
+enum class Dokumentstatus {
+    FERDIGSTILT,
+    AVBRUTT,
+    UNDER_REDIGERING,
+    KASSERT
+}
