+++ felles/src/main/kotlin/no/nav/familie/kontrakter/felles/journalpost/Dokumentvariant.kt
@@ -1,0 +1,3 @@
+package no.nav.familie.kontrakter.felles.journalpost
+
+data class Dokumentvariant(val variantformat: String)
