+++ felles/src/main/kotlin/no/nav/familie/kontrakter/felles/journalpost/Journalstatus.kt
@@ -1,0 +1,16 @@
+package no.nav.familie.kontrakter.felles.journalpost
+
+enum class Journalstatus {
+    MOTTATT,
+    JOURNALFOERT,
+    FERDIGSTILT,
+    EKSPEDERT,
+    UNDER_ARBEID,
+    FEILREGISTRERT,
+    UTGAAR,
+    AVBRUTT,
+    UKJENT_BRUKER,
+    RESERVERT,
+    OPPLASTING_DOKUMENT,
+    UKJENT
+}
