+++ felles/src/main/kotlin/no/nav/familie/kontrakter/felles/journalpost/LogiskVedlegg.kt
@@ -1,0 +1,3 @@
+package no.nav.familie.kontrakter.felles.journalpost
+
+data class LogiskVedlegg(val logiskVedleggId: String, val tittel: String)
