+++ felles/src/main/kotlin/no/nav/familie/kontrakter/felles/journalpost/RelevantDato.kt
@@ -1,0 +1,5 @@
+package no.nav.familie.kontrakter.felles.journalpost
+
+import java.time.LocalDateTime
+
+data class RelevantDato(val dato: LocalDateTime, val datotype: String)
