+++ felles/src/main/kotlin/no/nav/familie/kontrakter/felles/journalpost/Tema.kt
@@ -1,0 +1,7 @@
+package no.nav.familie.kontrakter.felles.journalpost
+
+enum class Tema {
+    BAR,
+    ENF,
+    KON
+}
